# One-line contextual logging of Kore actions