--- a/booster/tools/booster/Server.hs
+++ b/booster/tools/booster/Server.hs
@@ -30,7 +30,7 @@ import Data.InternedText (globalInternedTextCache)
 import Data.List (intercalate)
 import Data.List.Extra (splitOn)
 import Data.Map qualified as Map
-import Data.Maybe (fromMaybe, isJust, mapMaybe)
+import Data.Maybe (fromMaybe, isJust, mapMaybe, maybeToList)
 import Data.Set qualified as Set
 import Data.Text qualified as Text
 import Data.Text.Encoding qualified as Text (decodeUtf8)
@@ -161,7 +161,10 @@ main = do
             let koreLogActions :: forall m. MonadIO m => [LogAction m Log.SomeEntry]
                 koreLogActions = [koreStandardPrettyLogAction, koreJsonLogAction]
                   where
-                    logAsJson entry = Log.entryTypeText entry == "DebugAttemptEquation"
+                    logAsJson =
+                        if (Logger.LevelOther "SimplifyJson") `elem` customLevels
+                            then \entry -> Log.entryTypeText entry `elem` getKoreEntriesForLevel (Logger.LevelOther "SimplifyJson")
+                            else const False
 
                     koreStandardPrettyLogAction =
                         koreSomeEntryLogAction
@@ -324,6 +327,9 @@ logLevelToKoreLogEntryMap =
         , (LevelOther "SimplifySuccess", ["DebugApplyEquation"])
         , (LevelOther "RewriteSuccess", ["DebugAppliedRewriteRules"])
         ]
+
+getKoreEntriesForLevel :: LogLevel -> [Text.Text]
+getKoreEntriesForLevel level = concat . maybeToList $ Map.lookup level logLevelToKoreLogEntryMap
 
 data CLProxyOptions = CLProxyOptions
     { clOptions :: CLOptions