--- a/booster/tools/booster/Server.hs
+++ b/booster/tools/booster/Server.hs
@@ -32,12 +32,12 @@ import Data.List.Extra (splitOn)
 import Data.Map qualified as Map
 import Data.Maybe (fromMaybe, isJust, mapMaybe)
 import Data.Set qualified as Set
-import Data.Text (Text)
 import Data.Text qualified as Text
 import Data.Text.Encoding qualified as Text (decodeUtf8)
 import Options.Applicative
 import System.Clock (
     Clock (..),
+    TimeSpec (..),
     getTime,
  )
 import System.Exit
@@ -78,6 +78,9 @@ import Kore.Log.BoosterAdaptor (
     LogBoosterActionData (..),
     TimestampsSwitch (TimestampsDisable),
     defaultKoreLogOptions,
+    koreSomeEntryLogAction,
+    renderJson,
+    renderStandardPretty,
     swappableLogger,
     withLogger,
  )
@@ -155,33 +158,27 @@ main = do
 
             monadLogger <- askLoggerIO
 
-            koreLogEntriesAsJsonSelector <-
-                if Logger.LevelOther "SimplifyJson" `elem` customLevels
-                    then case Map.lookup (Logger.LevelOther "SimplifyJson") logLevelToKoreLogEntryMap of
-                        Nothing -> do
-                            Logger.logWarnNS
-                                "proxy"
-                                "Could not find out which Kore log entries correspond to the SimplifyJson level"
-                            pure (const 0)
-                        Just koreSimplificationLogEntries ->
-                            pure
-                                ( \x ->
-                                    if x `elem` koreSimplificationLogEntries
-                                        then 1
-                                        else 0
-                                )
-                    else pure (const 0)
-
-            let koreLogActions :: forall m. MonadIO m => [LogAction m Text]
+            let koreLogActions :: forall m. MonadIO m => [LogAction m Log.SomeEntry]
                 koreLogActions = [koreStandardPrettyLogAction, koreJsonLogAction]
                   where
-                    koreStandardPrettyLogAction = LogAction $ \txt ->
-                        liftIO $ monadLogger defaultLoc "kore" logLevel $ toLogStr txt
-                    koreJsonLogAction = LogAction $ \txt ->
-                        let bytes = Text.encodeUtf8 $ "[SimplifyJson] " <> txt <> "\n"
-                         in liftIO $ do
-                                BS.hPutStr simplificationLogHandle bytes
-                                IO.hFlush simplificationLogHandle
+                    koreStandardPrettyLogAction =
+                        koreSomeEntryLogAction
+                            (renderStandardPretty (ExeName "") (TimeSpec 0 0) TimestampsDisable)
+                            (not . (`elem` ["DebugAttemptEquation"]))
+                            (LogAction $ \txt -> liftIO $ monadLogger defaultLoc "kore" logLevel $ toLogStr txt)
+
+                    koreJsonLogAction =
+                        koreSomeEntryLogAction
+                            (renderJson (ExeName "") (TimeSpec 0 0) TimestampsDisable)
+                            (`elem` ["DebugAttemptEquation"])
+                            ( LogAction $ \txt ->
+                                let bytes = Text.encodeUtf8 $ prefix txt <> "\n"
+                                 in liftIO $ do
+                                        BS.hPutStr simplificationLogHandle bytes
+                                        IO.hFlush simplificationLogHandle
+                            )
+                      where
+                        prefix = if simplificationLogHandle == IO.stderr then ("[SimplifyJson] " <>) else id
 
             liftIO $ void $ withBugReport (ExeName "kore-rpc-booster") BugReportOnError $ \_reportDirectory -> withMDLib llvmLibraryFile $ \mdl -> do
                 let coLogLevel = fromMaybe Log.Info $ toSeverity logLevel
@@ -196,7 +193,6 @@ main = do
                                 LogBooster $
                                     Log.LogBoosterActionData
                                         { messageFilter = const True
-                                        , messageLogActionIndex = koreLogEntriesAsJsonSelector
                                         , logActions = koreLogActions
                                         }
                             , Log.logFormat = Log.Standard