… dev-tools/kore-rpc-dev/Server.hs
… kore-rpc-types/kore-rpc-types.cabal
… kore-rpc-types/src/Kore/Util.hs
… kore/app/rpc/Main.hs
… kore/src/Kore/Equation/DebugEquation.hs
… kore/src/Kore/JsonRpc.hs
… kore/src/Kore/Log.hs
… kore/src/Kore/Log/BoosterAdaptor.hs
… kore/src/Kore/Log/DebugAppliedRewriteRules.hs
… kore/src/Kore/Log/DebugAttemptUnification.hs
… kore/src/Kore/Log/DebugAttemptedRewriteRules.hs
… kore/src/Kore/Log/DebugContext.hs
… kore/src/Kore/Log/KoreLogOptions.hs
… kore/src/Kore/Log/Registry.hs
… kore/src/Kore/Repl/State.hs
… kore/src/Kore/Rewrite/Step.hs
… kore/src/Log/Entry.hs
… kore/test/Test/Kore/Rewrite/RewriteStep.hs
… scripts/booster-integration-tests.sh