@geo2a 2024-05-07
Kontrol performance with 1 worker: 

| Test                                      | georgy-better-kore-loggin time | master-88ef647ec time | (georgy-better-kore-loggin/master-88ef647ec) time
|-------------------------------------------|--------------------------------|-----------------------|---------------------------------------------------
| BlockParamsTest.testCoinBase()            | 39.25                          | 41.39                 | 0.9482966900217443
| CoinBaseTest.test_coinbase_setup()        | 69.0                           | 72.68                 | 0.9493670886075949
| BytesTypeTest.test_bytes4(bytes4)         | 36.89                          | 38.51                 | 0.9579330044144378
| FreshCheatcodes.test_address()            | 69.19                          | 66.27                 | 1.0440621699109702
| MethodDisambiguateTest.test_method_call() | 34.94                          | 31.32                 | 1.115581098339719
| TOTAL                                     | 249.26999999999998             | 250.17000000000002    | 0.996402446336491